# Optimize memory consumption in ResultCacheManager